refactor: `infrastructure` 하위 클래스들을 도메인별 패키지로 재구성
- `auth` 및 `sms` 관련 클래스들을 각각 `fittoring.infrastructure.auth`와 `fittoring.infrastructure.sms`로 이동
- 모든 참조 경로를 변경된 패키지 경로로 수정
- 테스트 코드 및 서비스 로직에서 새로운 경로 반영
