+++ backend/fittoring/src/main/java/fittoring/infrastructure/auth/CodeGenerator.java
@@ -1,4 +1,4 @@
-package fittoring.infrastructure;
+package fittoring.infrastructure.auth;
 
 public interface CodeGenerator {
