+++ backend/fittoring/src/test/java/fittoring/infrastructure/CodeGeneratorStub.java
@@ -1,5 +1,6 @@
 package fittoring.infrastructure;
 
+import fittoring.infrastructure.auth.CodeGenerator;
 import org.springframework.stereotype.Component;
 
 @Component
