+++ backend/fittoring/src/test/java/fittoring/integration/SmsOutboxDiscordNotificationIntegrationTest.java
@@ -13,8 +13,8 @@ import fittoring.domain.model.Phone;
 import fittoring.application.reservation.sms.SmsOutbox;
 import fittoring.application.reservation.sms.SmsOutboxEventType;
 import fittoring.application.reservation.sms.SmsOutboxStatus;
-import fittoring.infrastructure.SmsOutboxPublisher;
-import fittoring.infrastructure.SmsOutboxResultApplier;
+import fittoring.infrastructure.sms.SmsOutboxPublisher;
+import fittoring.infrastructure.sms.SmsOutboxResultApplier;
 import fittoring.infrastructure.exception.InfraErrorMessage;
 import fittoring.infrastructure.exception.SmsException;
 import org.assertj.core.api.SoftAssertions;
