+++ backend/fittoring/src/test/java/fittoring/integration/SmsOutboxPublisherIntegrationTest.java
@@ -11,8 +11,8 @@ import fittoring.domain.model.Phone;
 import fittoring.application.reservation.sms.SmsOutbox;
 import fittoring.application.reservation.sms.SmsOutboxEventType;
 import fittoring.application.reservation.sms.SmsOutboxStatus;
-import fittoring.infrastructure.SmsOutboxPublisher;
-import fittoring.infrastructure.SmsOutboxResultApplier;
+import fittoring.infrastructure.sms.SmsOutboxPublisher;
+import fittoring.infrastructure.sms.SmsOutboxResultApplier;
 import fittoring.infrastructure.dto.BatchSendResult;
 import fittoring.infrastructure.exception.InfraErrorMessage;
 import fittoring.infrastructure.exception.SmsException;
